finalizando crud

+++ src/main/java/com/example/demo/repository/AtendimentoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Atendimento;
 
-public interface AtendimentoRepository extends JpaRepository<Atendimento, Long>{ 
+public interface AtendimentoRepository extends JpaRepository<Atendimento, Integer>{ 
 
 
 }
